fix: Fix Recipe models data type

+++ src/main/java/mg/sandratra/bakery/models/recipe/RecipeIngredient.java
@@ -10,8 +10,8 @@ import lombok.NoArgsConstructor;
 @NoArgsConstructor
 @AllArgsConstructor
 public class RecipeIngredient {
-    private int id;
-    private int recipeId;
-    private int ingredientId;
+    private Long id;
+    private Long recipeId;
+    private Long ingredientId;
     private BigDecimal quantity;
 }
